+++ src/index.ts
@@ -1,14 +1,14 @@
 import { Touch } from './instance'
-import { EventHub } from './utils/event'
-import { MouseHub } from './utils/mouse'
-
+import { EventHub } from './pool/event'
+import { MouseHub } from './pool/mouse'
+import { Card } from './pool/card'
 
 ;(async() => {
   if (!document) return
   const hub: EventHub = new EventHub()
   new MouseHub(hub)
+  new Card(hub)
   new Touch(hub)
-  
   
 })()
