+++ src/instance.ts
@@ -1,4 +1,5 @@
-import { EventHub } from './utils/event'
+import { EventHub } from './pool/event'
+import { toEnglish } from './core/translator'
 
 
 export class Touch {
@@ -6,8 +7,10 @@ export class Touch {
   
   constructor(eventHub: EventHub) {
     this.hub = eventHub
-    this.hub.listen('updateText', (e, text) => {
+    this.hub.listen('updateText', (e, { text, position }) => {
       console.log(text)
+      // toEnglish(text).then()
+      this.hub.dispath('showCard', { dist: '一段文字', position })
     })
   }
