+++ src/utils/page.ts
@@ -2,3 +2,7 @@
 export const findSelectdText = (): string => {
   return `${window.getSelection()}`
 }
+
+
+
+
